+++ CHANGELOG.md
@@ -19,10 +19,13 @@
 - Fixed pcsd crash when rpam rubygem is installed ([ghissue#109])
 - Fixed occasional crashes / failures when using locale other than en_US.UTF8
   ([rhbz#1387106])
+- Fixed starting and stopping cluster services on systemd machines without
+  the `service` executable ([ghissue#115])
 
 
 [ghissue#81]: https://github.com/ClusterLabs/pcs/issues/81
 [ghissue#109]: https://github.com/ClusterLabs/pcs/issues/109
+[ghissue#115]: https://github.com/ClusterLabs/pcs/issues/115
 [rhbz#1262001]: https://bugzilla.redhat.com/show_bug.cgi?id=1262001
 [rhbz#1380372]: https://bugzilla.redhat.com/show_bug.cgi?id=1380372
 [rhbz#1387106]: https://bugzilla.redhat.com/show_bug.cgi?id=1387106